requirements.txt
# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.4

# Scientific Computing
numpy>=1.19.0
scipy>=1.7.0
scikit-image>=0.18.0

# Face Detection & Alignment
face-alignment>=1.3.5

# 3D Mesh Processing
trimesh>=3.9.0
open3d>=0.13.0

# Data Processing
pandas>=1.3.0
scikit-learn>=0.24.0

# Visualization (optional)
matplotlib>=3.4.0
plotly>=5.0.0

# Development (optional)
pytest>=6.2.0
